ambedkargpt/requirements.txt
# Core NLP & Embeddings
sentence-transformers>=2.2.0
spacy>=3.7.0
nltk>=3.8.0
transformers>=4.35.0

# Knowledge Graph & Graph Processing
networkx>=3.2.0
python-louvain>=0.16
leidenalg>=0.10.0
igraph>=0.11.0

# Vector Storage & Similarity
chromadb>=0.4.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0

# LLM Integration
ollama>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.10

# PDF Processing
pypdf>=3.17.0
pdfplumber>=0.10.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0
rich>=13.7.0

# Testing (optional)
pytest>=7.4.0

# Streamlit UI
streamlit>=1.29.0
plotly>=5.18.0

